+++ index.html
@@ -31,6 +31,5 @@
     <div id="js-game"></div>
     <p id="js-score" class="score"></p>
     <script src="./src/scripts/main.js" type="module"></script>
-    <script src="./src/scripts/explanation.js" type="module"></script>
   </body>
 </html>
