--- src/scripts/config/board.js
@@ -1,9 +1,0 @@
-/**
- * canvasの大きさや行と列の設定
- */
-export const boardConfig = {
-  H: 600,
-  W: 300,
-  ROWS: 20,
-  COLS: 10,
-};
