+++ src/scripts/controllers/index.js
@@ -1,0 +1,1 @@
+export * from "./game";
