+++ src/scripts/views/index.js
@@ -2,3 +2,4 @@ export * from "./block";
 export * from "./board";
 export * from "./canvas";
 export * from "./score";
+export * from "./gameHandler";
